webidl: track stability of individual interface & namespace members (#3214)
Fixes #2560

It's possible for some members of an interface (or namespace) to be stable and others not, if a `partial interface` is defined in an unstable webidl file which adds to an interface defined in a stable file. This wasn't properly kept track of for all interface members before, most notebly for operations. This PR fixes that.