Add missing guide pages to `SUMMARY.md` (#3521)
There were several pages that had been added to the guide, but were inaccessible due to not being listed in `SUMMARY.md` (which is how you specify the layout for the left sidebar). This PR fixes that.

I also noticed that there were two pages for the same `typscript_type` attribute: one in the 'on JS imports' section (which was inaccessible) and one in the 'on Rust exports' section. These aren't for two different versions of the attribute; they're both for the same attribute which can only be applied to JS imports.

The one in the 'on JS imports' section was much less detailed, so I got rid of it and moved the other one into its place.

Closes #2426 (an issue that was opened ages ago about the `typescript_type` page being in the wrong section). There was a bit of discussion there about whether it does actually belong in the 'on Rust exports' section, since despite being directly applied to JS imports its usage is to change the TypeScript output for Rust exports, but I think that the site where it's applied is more important.