Fix flaky `TypedArray::from(&[T])`
Instead of going via an intermediate unsafe view, we can add another safe binding that takes a slice "directly" and clones it into a typed array.

This is safer, generates slightly compact JS with fewer roundtrips, and fixes the flakiness due to allocations that used to happen between creating the slice and actually cloning it.

Fixes #4492.