Self-host almost all intrinsic signatures (#4636)
* Self-host almost all intrinsic signatures

Instead of describing intrinsic signatures out of band in cli-support, manually describing their raw signatures in wasm-bindgen, and unsafely converting from one to another, we can just declare almost all the intrinsics as a module import via the `#[wasm_bindgen]` macro.

This way they're guaranteed to stay in sync, and we can get rid of a bunch of `unsafe` code.

* Clippy fixup

* Limit unsafe scope of Closure::drop